Refactor MainScreen to fetch weather data based on selected units
The MainScreen composable has been updated to retrieve the selected unit (metric or imperial) from the UnitsViewModel.
The weather data is now fetched using the specified unit and displayed accordingly.
A default city "Ouargla" is set if no city is provided.
The unit is now saved in a state variable.
Added Log.d to show the weather description.